set up router and dashboard redirect on reg
reg syntax naming is not really appropriate.
routes for sign in vs register are not really great.
not accessing user data to front end after reg.
First, need to sort out sign in and reg distinction and clean up code naming and logiv
second, need to implement proper sign in and user data access for frontend.